Move CLI command tests to integration tests
