--- radicle-cli/src/tests.rs
@@ -1,1 +1,0 @@
-mod cli;
